lab/06_pgvector_indexes/requirements.txt
psycopg2-binary
pgvector
openai
python-dotenv
transformers
torch
sentencepiece
numpy
